+++ client/src/features/gamemaster/components/Sidebar.tsx
@@ -1,0 +1,139 @@
+import React, { useState, useEffect } from "react";
+import styled from "styled-components";
+import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
+import { Button } from "components";
+import { Act } from "../types";
+
+const reorder = (list: any[], startIndex: number, endIndex: number) => {
+  const result = Array.from(list);
+  const [removed] = result.splice(startIndex, 1);
+  result.splice(endIndex, 0, removed);
+
+  return result;
+};
+
+type Props = {
+  acts: Act[];
+  setActs: (_acts: Act[]) => void;
+  selectedAct: Act;
+  setSelectedAct: (_act: Act) => void;
+};
+
+export const Sidebar: React.FC<Props> = ({
+  acts,
+  setActs,
+  selectedAct,
+  setSelectedAct,
+}) => {
+  const onDragEnd = (result: any) => {
+    if (!result.destination) {
+      return;
+    }
+
+    const orderedActs = reorder(
+      acts,
+      result.source.index,
+      result.destination.index
+    );
+
+    setActs(orderedActs);
+  };
+
+  const onSelectAct = (act: Act) => {
+    setSelectedAct(act);
+  };
+
+  return (
+    <SidebarContainer>
+      <SidebarHeader>
+        <h3>Questions:</h3>
+      </SidebarHeader>
+      <SidebarContent>
+        <DragDropContext onDragEnd={onDragEnd}>
+          <Droppable droppableId="droppable">
+            {(provided) => (
+              <SidebarContent
+                {...provided.droppableProps}
+                ref={provided.innerRef}
+              >
+                {acts.map((act, index) => (
+                  <Draggable key={act.id} draggableId={act.id} index={index}>
+                    {(provided) => (
+                      <QuestionItem
+                        ref={provided.innerRef}
+                        {...provided.draggableProps}
+                        {...provided.dragHandleProps}
+                        isSelected={selectedAct?.id === act.id}
+                        style={{ ...provided.draggableProps.style }}
+                        onClick={() => onSelectAct(act)}
+                      >
+                        <div>
+                          <div className="instruction">{act.instruction}</div>
+                          <h3 className="question">{act.question}</h3>
+                        </div>
+                        <div className="type">{act.questionType}</div>
+                      </QuestionItem>
+                    )}
+                  </Draggable>
+                ))}
+                {provided.placeholder}
+              </SidebarContent>
+            )}
+          </Droppable>
+        </DragDropContext>
+      </SidebarContent>
+      <SidebarFooter>
+        <Button>Add new question</Button>
+      </SidebarFooter>
+    </SidebarContainer>
+  );
+};
+
+const SidebarContainer = styled.section`
+  grid-area: sidebar;
+  background: ${({ theme }) => theme.ui.background};
+  padding: 0 ${({ theme }) => theme.spacings(3)};
+  display: grid;
+  grid-template-rows: max-content auto max-content;
+  height: 100%;
+`;
+
+const SidebarHeader = styled.header``;
+
+const SidebarContent = styled.section`
+  overflow: auto;
+`;
+
+const SidebarFooter = styled.footer`
+  padding: ${({ theme }) => theme.spacings(3)} 0;
+`;
+
+const QuestionItem = styled.div<{ isSelected: boolean }>`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  padding: ${({ theme }) => theme.spacings(3)};
+  border: 2px dotted ${({ theme }) => theme.colors.lightGrey};
+  border-radius: ${({ theme }) => theme.border.wavyRadius};
+  margin-bottom: ${({ theme }) => theme.spacings(3)};
+  background-color: ${({ theme, isSelected }) =>
+    isSelected ? theme.ui.backgroundGrey : theme.ui.background};
+
+  .instruction {
+    color: ${({ theme }) => theme.colors.darkGrey};
+    font-size: 14px;
+    margin-bottom: ${({ theme }) => theme.spacings(2)};
+  }
+
+  .question {
+    font-size: 24px;
+    margin: 0 0 ${({ theme }) => theme.spacings(3)};
+  }
+
+  .type {
+    height: fit-content;
+    padding: ${({ theme }) => theme.spacings(1)};
+    border: 2px solid ${({ theme }) => theme.ui.modal.border};
+    border-radius: ${({ theme }) => theme.border.wavyRadius};
+  }
+`;
