+++ client/src/features/gamemaster/types.ts
@@ -1,0 +1,7 @@
+export type Act = {
+  id: string;
+  questionType: string;
+  instruction: string;
+  question: string;
+  answerType: string;
+};
